Dataset summarizing 1000 posterior samples of BAMM for extant ponerine ants updated to 10 My

An object of class "bammdata" containing information of diversification dynamics of extant ponerine ants (Ponerinae subfamily) modeled with BAMM and cut-off for a focal_time of 10 My. This object was obtained with update_rates_and_regimes_for_focal_time().

Source: Doré, M., Borowiec, M. L., Branstetter, M. G., Camacho, G. P., Fisher, B. L., Longino, J. T., Ward, P. S., Blaimer, B. B., (2025), Evolutionary history of ponerine ants highlights how the timing of dispersal events shapes modern biodiversity, Nature Communications. https://doi_of_Paper_to_provide.html

Usage

data(Ponerinae_BAMM_object_10My)

Format

A list with 32 elements.

Details

An object of class "bammdata" containing information of diversification dynamics of extant ponerine ants (Ponerinae subfamily) modeled with BAMM.

Phylogeny-related elements used to plot a phylogeny with ape::plot.phylo():

$edge Matrix of integers. Defines the tree topology by providing rootward and tipward node ID of each edge.

$Nnode Integer. Number of internal nodes.

$tip.label Vector of character strings. Labels of all tips. If an initial extant clade was cut-off, the tip.label is the tipward edge ID of the cut branch.

$edge.length Vector of numerical. Length of edges/branches.

BAMM internal elements used for tree exploration:

$begin Vector of numerical. Absolute time since root of edge/branch start (rootward).

$end Vector of numerical. Absolute time since root of edge/branch end (tipward).

$downseq Vector of integers. Order of node visits when using a pre-order tree traversal.

$lastvisit ID of the last node visited when starting from the node in the corresponding position in $downseq.

BAMM elements summarizing diversification data updated for a focal_time of 10 My:

$numberEvents Vector of integer. Number of events/macroevolutionary regimes (k+1) recorded in each posterior configuration. k = number of shifts.

$eventData List of data.frames. One per posterior sample. Records shift events and macroevolutionary regimes parameters. 1st line = Background root regime.

$eventVectors List of integer vectors. One per posterior sample. Record regime ID per branches.

$tipStates List of integer vectors. One per posterior sample. Record regime ID per tips.

$tipLambda List of numerical vectors. One per posterior sample. Record speciation rates per tips.

$tipMu List of numerical vectors. One per posterior sample. Record extinction rates per tips.

$eventBranchSegs List of matrix of numerical. One per posterior sample. Record regime ID per segments of branches.

$meanTipLambda Vector of numerical. Mean tip speciation rates across all posterior configurations of tips.

$meanTipMu Vector of numerical. Mean tip extinction rates across all posterior configurations of tips.

$type Character string. Set the type of data modeled with BAMM. Here, type = "diversification".

Additional elements providing key information for downstream analyses:

$expectedNumberOfShifts Integer. The expected number of regime shifts used to set the prior in BAMM.

$MSP_tree Object of class phylo. List of 4 elements duplicating information from the Phylogeny-related elements above, except $MSP_tree$edge.length is recording the Marginal Shift Probability of each branch (i.e., the probability of a regime shift to occur along each branch) whose origin is older that focal_time.

$MAP_indices Vector of integers. The indices of the Maximum A Posteriori probability (MAP) configurations among the posterior samples.

$MAP_BAMM_object. List of 18 elements of class "bammdata" recording the mean rates and regime shift locations found across the Maximum A Posteriori probability (MAP) configuration. All BAMM elements summarizing diversification data holds a single entry describing this the mean diversification history, updated for the focal_time`.

$MSC_indices Vector of integers. The indices of the Maximum Shift Credibility (MSC) configurations among the posterior samples.

$MSC_BAMM_object List of 18 elements of class "bammdata" recording the mean rates and regime shift locations found across the Maximum Shift Credibility (MSC) configurations. All BAMM elements summarizing diversification data holds a single entry describing this mean diversification history, updated for the focal_time`.

New elements added to provide updated information:

$root_age Integer. Stores the age of the root of the tree.

$nodes_ID_df Data.frame with two columns. Provides the conversion from the new_node_ID in the cut tree to the initial_node_ID in the extant tree. Each row is a node.

$initial_nodes_ID Vector of character strings. Provides the initial ID of internal nodes. Used to plot internal node IDs as labels with ape::nodelabels().

$edges_ID_df Data.frame with two columns. Provides the conversion from the new_edge_ID in the cut tree to the initial_edge_ID in the extant tree. Each row is an edge/branch.

$initial_edges_ID Vector of character strings. Provides the initial ID of edges/branches. Used to plot edge/branch IDs as labels with ape::edgelabels().

$dtrates List of three elements.

1/ $dtrates$tau Numerical. Resolution factor describing the fraction of each segment length used in plot_BAMM_rates() compared to the full depth of the initial tree (i.e., the root_age)

2/ $dtrates$rates List of two numerical vectors. Speciation and extinction rates along segments used by plot_BAMM_rates().

3/ $dtrates$tmat Matrix of numerical. Start and end times of segments in term of distance to the root.

$initial_colorbreaks List of three vectors of numerical. Rate values of the percentiles delimiting the bins for mapping rates to colors with BAMMtools::plot.bammdata(). Each element provides values for different type of rates ($speciation, $extinction, $net_diversification).

$focal_time Integer. The time, in terms of time distance from the present, at which the rates/regimes were extracted and the tree was eventually cut. Here: 10 My.

References

Doré, M., Borowiec, M. L., Branstetter, M. G., Camacho, G. P., Fisher, B. L., Longino, J. T., Ward, P. S., Blaimer, B. B., (2025), Evolutionary history of ponerine ants highlights how the timing of dispersal events shapes modern biodiversity, Nature Communications. https://doi.org/10.1038/s41467-025-63709-3

See also